ruc_dni tipo Long

+++ ClientJPA/src/main/java/com/accountbank/service/ClientService.java
@@ -7,8 +7,8 @@ import java.util.Optional;
 
 public interface ClientService {
     List<Client> getClients();
-    List<Client> getById(Integer ruc_dni);
+    List<Client> getById(Long ruc_dni);
     List<Client> create(Client client);
     List<Client> update(Client client);
-    void delete(Integer ruc_dni);
+    void delete(Long ruc_dni);
 }
